+++ CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Loading Spinners
 - Footer
 
+### Fixed
+
+- get Location (call n+1)
+
 ## [0.1.0] - 2020-04-15
 
 ### Added